refactor(dashboard): extract secondary risk dedupe behind size ratchet
Move the duplicate-review detection cluster (secondary risk summary,
stopped command text, envelope display text) into secondary-risk-dedupe
and re-export it from approval-center-utils so existing imports keep
working. approval-center-utils and phase09-review.test.ts sit on the
oversized-file ratchet, so the compound-findings regression tests move
to a dedicated secondary-risk-dedupe test file wired into the test
chain.
